requirements.txt
grpcio~=1.50.0
grpcio-tools~=1.50.0
redis~=4.3.4
psutil~=5.9.3
flask~=2.2.2
speedtest-cli~=2.1.3
requests~=2.28.1
urllib3~=1.26.6